+++ src/main/java/com/shinesolutions/aemorchestrator/exception/NoPairFoundException.java
@@ -1,0 +1,10 @@
+package com.shinesolutions.aemorchestrator.exception;
+
+public class NoPairFoundException extends Exception {
+    
+    private static final long serialVersionUID = 1L;
+
+    public NoPairFoundException(String instanceId) {
+        super("Unable to find available pair for instance " + instanceId);
+    }
+}