+++ src/main/java/com/shinesolutions/aemorchestrator/service/AemInstanceHelperService.java
@@ -24,6 +24,8 @@ import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Component;
 
 import com.shinesolutions.aemorchestrator.exception.InstanceNotInHealthyState;
+import com.shinesolutions.aemorchestrator.exception.NoPairFoundException;
+import com.shinesolutions.aemorchestrator.model.EC2Instance;
 import com.shinesolutions.aemorchestrator.model.EnvironmentValues;
 import com.shinesolutions.aemorchestrator.model.InstanceTags;
 import com.shinesolutions.aemorchestrator.util.HttpUtil;
@@ -311,11 +313,29 @@ public class AemInstanceHelperService {
      * @return Publish Dispatcher instance ID tag
      * @throws NoSuchElementException if can't find unpaired Publish Dispatcher
      */
-    @Retryable(maxAttempts=10, value=NoSuchElementException.class, backoff=@Backoff(delay=5000))
-    public String findUnpairedPublishDispatcher(String instanceId) throws NoSuchElementException {
-        List<String> dispatcherIds = awsHelperService.getInstanceIdsForAutoScalingGroup(
+    @Retryable(maxAttempts=10, value=NoPairFoundException.class, backoff=@Backoff(delay=5000))
+    public String findUnpairedPublishDispatcher(String instanceId) throws NoPairFoundException {
+        String unpairedDispatcher = null;
+        
+        List<EC2Instance> dispatcherIds = awsHelperService.getInstancesForAutoScalingGroup(
             envValues.getAutoScaleGroupNameForPublishDispatcher());
-        return dispatcherIds.stream().filter(d -> isViablePair(instanceId, d)).findFirst().get();
+        //Filter the list to get all possible eligible candidates
+        dispatcherIds = dispatcherIds.stream().filter(d -> 
+            isViablePair(instanceId, d.getInstanceId())).collect(Collectors.toList());
+        
+        if(dispatcherIds.size() > 1) {
+            String publishAZ = awsHelperService.getAvailabilityZone(instanceId);
+            
+            //If there are many candidates, then pick the one with the same AZ or else use first
+            unpairedDispatcher = (dispatcherIds.stream().filter(i -> i.getAvailabilityZone().equalsIgnoreCase(publishAZ))
+                .findFirst().orElse(dispatcherIds.get(0))).getInstanceId();
+        } else if (dispatcherIds.size() == 1) {
+            unpairedDispatcher = dispatcherIds.get(0).getInstanceId();
+        } else {
+            throw new NoPairFoundException(instanceId);
+        }
+        
+        return unpairedDispatcher;
     }
     
     private boolean isViablePair(String instanceId, String dispatcherInstanceId) {