+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Now reads the SNS topic ARN from Cloud Formation based on the logical id #21
 * Scale down handlers now cater for inexisting replication/flush agent #18
 * Creation of reverse replication agents now includes 'userId' attribute
+* Prioritises Publish and Publish-Dispatcher instances from same AZ #3
 
 ### 0.9.2
 * Added configurable snapshot tags #2